fix: websocket connection

+++ lib/features/module_list/domain/data/repositories/module_list_repository.dart
@@ -7,7 +7,7 @@ class ModuleListRepository {
   Future<List<ModuleModel>> getModulesByLocation({
     required double latitude,
     required double longitude,
-    required int radius,
+    required double radius,
   }) async {
     try {
       final response = await _dio.instance.post(
